lxd: Move storage cgo to storage package
